+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - feat(pipeline): add `Waveform` and `SampleRate` preprocessors
 - feat(model): add `num_frames`, `receptive_field`, and `dimension` properties to segmentation models
 - feat(sample): add sample file at `pyannote.audio.sample.SAMPLE_FILE`
+- feat(powerset): add `Powerset.permutation_mapping` to help with permutation in powerset space
 
 ### Fixes